+++ src/main/java/net/earomc/synthesizer/waveform/waveforms/Noise.java
@@ -6,7 +6,7 @@ import java.util.concurrent.ThreadLocalRandom;
 
 public class Noise implements Waveform {
     @Override
-    public float sample(float timeSeconds, float periodSeconds, float amp, float phaseRadians) {
+    public float sample(float timeSeconds, double periodSeconds, float amp, float phaseRadians) {
         return amp * ThreadLocalRandom.current().nextFloat(-1f, 1f);
     }
 }
