+++ src/main/java/net/earomc/synthesizer/waveform/waveforms/Square.java
@@ -4,7 +4,7 @@ import net.earomc.synthesizer.waveform.Waveform;
 
 public class Square implements Waveform {
     @Override
-    public float sample(float timeSeconds, float periodSeconds, float amp, float phaseRadians) {
+    public float sample(float timeSeconds, double periodSeconds, float amp, float phaseRadians) {
         return (float) (amp * Math.signum(Math.sin((2 * Math.PI * timeSeconds - phaseRadians) / periodSeconds)));
     }
 }
